v0.13.3: zero clippy warnings, rustfmt everywhere, CI actually green
Hygiene + CI-reliability release. No user-facing feature changes.

Lint (workspace at zero, gated):
- ~55 clippy warnings across all targets -> 0
- live_stress SERIAL mutex -> tokio::sync::Mutex (guard held across test
  bodies; std guard blocked the runtime worker; also drops poisoning)
- render_space 12 positional args -> 5 via SpaceSocialContext +
  SpaceRenderPrefs structs
- handle_update_role PATCH fields -> RoleUpdate struct
- create_scheduled_event / schedule_message -> NewScheduledEvent /
  NewScheduledMessage param structs; get_due_scheduled_messages 6-tuple ->
  DueScheduledMessage row struct
- ChunkedScreenSequenceState::{NewFrame,ExistingFrame,StaleFrame} ->
  {New,Existing,Stale}; SoundboardCombos type alias; while_let /
  needless_range_loop / iter_nth / field_reassign sweeps
- CI clippy gate tightened from "<=63 warnings" ratchet to -D warnings
- cargo fmt --all run once (46 files); fmt --check gate now passes

CI (fixed the two structural failures keeping every run red):
- build-test: server_tests spawn target/debug/{signaling_server,app_desktop}
  which cargo test never builds on a fresh runner -> added explicit
  cargo build step
- windows-installer: vc_redist.x64.exe is gitignored and ci.yml never
  downloaded it -> added the release.yml download step

Real bug surfaced by the CI fix:
- The wave-12 bounded media lane (capacity 8) was never actually
  integration-tested locally because target/debug/signaling_server was
  stale. A same-host 100-frame burst dropped ~2/3 of frames deterministically.
  MEDIA_LANE_CAPACITY 8 -> 32 (absorbs TCP delivery bunching after network
  hiccups; ~640ms audio headroom; frames are shared Bytes so a full queue
  holds refcounts, not copies). test_audio_relay_stress now paces frames at
  2ms like a real client instead of asserting on unguaranteed burst behavior.

Also in this release (landed alongside):
- Role-assignment privilege-escalation guard (AssignRoleToMember verifies
  space ownership of the role + actor holds every permission it grants)
- Presence friendship-gated (WatchFriendPresence no longer a privacy oracle)
- Timeout rank checks (no timing out equal/higher-ranked members;
  target must be in the actor's space)
- Scheduled-event deletion space-scoped
- voxlink.iss #ifndef guard so the release workflow's /DMyAppVersion wins

Tests: 112/112 server integration green against the freshly-built binary;
all unit suites green; UI snapshot matrix green.

+++ crates/signaling_server/src/relay/audio.rs
@@ -1,11 +1,11 @@
-use std::sync::Arc;
-use std::sync::atomic::Ordering;
-use tokio::net::UdpSocket;
+use crate::metrics_server::ServerMetrics;
 use crate::types::{Peer, State};
 use crate::validation::atomic_rate_check;
-use crate::metrics_server::ServerMetrics;
 use crate::LIMITS;
 use shared_types;
+use std::sync::atomic::Ordering;
+use std::sync::Arc;
+use tokio::net::UdpSocket;
 
 type Metrics = Arc<ServerMetrics>;
 
@@ -44,8 +44,10 @@ pub(crate) async fn relay_audio(
     // Without this, a moderation timeout only suppresses chat sends; voice was still
     // relayed, which made timeouts feel half-applied.
     {
-        let now = crate::now_epoch_secs() as u64;
-        let until = peer.timeout_until.load(std::sync::atomic::Ordering::Relaxed);
+        let now = crate::now_epoch_secs();
+        let until = peer
+            .timeout_until
+            .load(std::sync::atomic::Ordering::Relaxed);
         if until > now {
             return;
         }
@@ -171,8 +173,10 @@ pub(crate) async fn relay_audio_udp(
     // Server-enforced timeout: drop audio while under moderation timeout (mirrors
     // the WebSocket relay above so UDP can't bypass moderation).
     {
-        let now = crate::now_epoch_secs() as u64;
-        let until = peer.timeout_until.load(std::sync::atomic::Ordering::Relaxed);
+        let now = crate::now_epoch_secs();
+        let until = peer
+            .timeout_until
+            .load(std::sync::atomic::Ordering::Relaxed);
         if until > now {
             return;
         }
@@ -183,7 +187,9 @@ pub(crate) async fn relay_audio_udp(
         &peer.audio_frame_count,
         LIMITS.max_audio_fps,
     ) {
-        metrics.udp_rate_limited_total.fetch_add(1, Ordering::Relaxed);
+        metrics
+            .udp_rate_limited_total
+            .fetch_add(1, Ordering::Relaxed);
         return;
     }
 
@@ -268,7 +274,9 @@ pub(crate) async fn relay_audio_udp(
         if let Some(addr) = udp_addr {
             // Send via UDP — fire-and-forget (UDP is unreliable by design)
             if let Err(_e) = udp_socket.send_to(frame, addr).await {
-                metrics.udp_send_failures_total.fetch_add(1, Ordering::Relaxed);
+                metrics
+                    .udp_send_failures_total
+                    .fetch_add(1, Ordering::Relaxed);
             }
             metrics.udp_frames_out_total.fetch_add(1, Ordering::Relaxed);
         } else {
